Payment available now lol

+++ server/models/seat.go
@@ -11,7 +11,7 @@ type Seat struct {
 	Number       int       `gorm:"column:number" json:"number"`
 	CreatedAt    time.Time `gorm:"column:created_at" json:"created_at"`
 	Available    bool      `gorm:"column:available" json:"available"`
-	Price        int       `gorm:"column:price" json:"price"`
+	Price        int64     `gorm:"column:price" json:"price"`
 	Type         string    `gorm:"column:type" json:"type"`
 
 	Status int `gorm:"-" json:"status"`
